% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDFRowsMean.R
\name{getDFRowsMean}
\alias{getDFRowsMean}
\title{Calculate the mean values for specified rows}
\usage{
getDFRowsMean(dataFrame, valsToCompute)
}
\arguments{
\item{dataFrame}{The dataframe to analyse.}

\item{valsToCompute}{The values to analyse in that dataframe.}
}
\value{

}
\description{
Calculate the mean values for specified rows
}
\examples{
getDFRowsMean(myDataFrame, c("height", "weight"))

}
