% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRightCellValue.R
\name{getRightCellVal}
\alias{getRightCellVal}
\title{Retrieve values in the next cell}
\usage{
getRightCellVal(
  dataFrameList,
  valuesToGet,
  selectedDataFrame = selectedSubjects
)
}
\arguments{
\item{dataFrameList}{List containing the dataframes to analyse.}

\item{valuesToGet}{The left cell that is known (you want to get the
value at it right)}

\item{selectedDataFrame}{Dataframes to analyse in the provided list.}
}
\value{

}
\description{
Retrieve values in the next cell
}
